Custom iterators for arrays
