+++ src/index.ts
@@ -1,7 +1,8 @@
 export * from './a-iterable';
 export * from './api';
+export * from './array';
 export * from './array-like-iterable';
-export * from './construction';
+export * from './none';
 export * from './objects';
 export * from './reverse';
 export * from './revertible-iterable';
