+++ src/none.spec.ts
@@ -1,0 +1,12 @@
+import { overNone } from './none';
+
+describe('overNone', () => {
+  it('has no elements', () => {
+    expect([...overNone()]).toEqual([]);
+  });
+  describe('reverse', () => {
+    it('returns the same instance', () => {
+      expect(overNone().reverse()).toBe(overNone());
+    });
+  });
+});
